# Optimize performance (~3.5X faster build when images are up-to-date)