--- a/setup.py
+++ b/setup.py
@@ -26,6 +26,7 @@ setuptools.setup(
         'Click==7.0',
         'pyaml==19.4.1',
         'docker==3.7.0',
+        'typing-extensions==3.7.4.3',
         'yamllint==1.17.0',
         "wcmatch==6.0.1"
     ]