This is a second iteration of #44, where we optimize the performance of Brick when images are already built. 

In #44 we used the modified timestamps of inputs and the BUILD.yaml file to speed up `brick prepare` and `brick build`. But as git does not preserve the timestamps, we **now rely on the hash of the content of all dependency files** (inputs and BUILD.yaml file). 

Note that hashing is more expensive that comparing modified timestamps (which we also didn't do correctly in #44, but only on folders). But for all the targets at tmrow the execution time of `sha1sum` for all dependency files is roughly 10-200 ms. At least locally this is considerably faster than the overhead of buildkit (the bottleneck here is "resolve image config" and "load metadata"), which is around 1-1.5 seconds per image when everything is cached.

## Design consideration

First I did a ~/.brick-cache.json file containing all hashes for tags. But I realized that this would become out of date with the docker images once images are pruned, etc. So instead we now simply add a `LABEL` (key/value part of the Dockerfile) to the image. So we have a single source of truth. And the lookup is simple: "do we have any images built for tag X and Y with dependency hash A?".

## Timings

I got these timings locally when nothing changed and everything was already prepared and built:

    time brick -r prepare    : 48 seconds 
    time brick -r build      : 175 seconds

On this branch:

    time brick -r prepare    : 5 seconds
    time brick -r build      : 52 seconds


So when everything is up-to-date: brick build time is ~3.5X faster and prepare time is 10X faster. 🚀


## TODO

- [x] Depends on merging https://github.com/tmrowco/brick/pull/45
- [x] Test on CI and verify performance increase