@madsnedergaard 2020-09-01
# MAGIC! 🧙 

<img width="1035" alt="Screenshot 2020-09-01 at 10 10 08" src="https://user-images.githubusercontent.com/3296643/91824615-55818a80-ec3b-11ea-9971-c7410e01084d.png">